+++ android/koin-android/src/main/java/org/koin/androidx/scope/LifecycleScopeDelegate.kt
@@ -9,28 +9,35 @@ import org.koin.core.component.getScopeId
 import org.koin.core.component.getScopeName
 import org.koin.core.context.GlobalContext
 import org.koin.core.context.KoinContext
+import org.koin.core.logger.Logger
 import org.koin.core.scope.Scope
 import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.KProperty
 
 class LifecycleScopeDelegate<T>(
-        val lifecycleOwner: LifecycleOwner,
-        private val koinContext: KoinContext = GlobalContext,
-        private val createScope: (Koin) -> Scope = { koin: Koin -> koin.createScope(lifecycleOwner.getScopeId(), lifecycleOwner.getScopeName(), lifecycleOwner as T) },
+    val lifecycleOwner: LifecycleOwner,
+    private val koin: Koin,
+    private val createScope: (Koin) -> Scope = { koin: Koin ->
+        koin.createScope(
+            lifecycleOwner.getScopeId(),
+            lifecycleOwner.getScopeName(),
+            lifecycleOwner as T
+        )
+    },
 ) : ReadOnlyProperty<LifecycleOwner, Scope> {
 
     private var _scope: Scope? = null
 
     init {
-        val koin = koinContext.get()
         val logger = koin.logger
 
         logger.debug("setup scope: $_scope for $lifecycleOwner")
-        val scopeId = lifecycleOwner.getScopeId()
-        _scope = koin.getScopeOrNull(scopeId) ?: createScope(koin)
-        logger.debug("got scope: $_scope for $lifecycleOwner")
-
         lifecycleOwner.lifecycle.addObserver(object : LifecycleObserver {
+            @OnLifecycleEvent(Lifecycle.Event.ON_CREATE)
+            fun onCreate(owner: LifecycleOwner) {
+                createScope()
+            }
+
             @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
             fun onDestroy(owner: LifecycleOwner) {
                 logger.debug("Closing scope: $_scope for $lifecycleOwner")
@@ -42,17 +49,21 @@ class LifecycleScopeDelegate<T>(
         })
     }
 
+    private fun createScope() {
+        if (_scope == null) {
+            koin.logger.debug("Create scope: $_scope for $lifecycleOwner")
+            val scopeId = lifecycleOwner.getScopeId()
+            _scope = koin.getScopeOrNull(scopeId) ?: createScope(koin)
+        }
+    }
+
     override fun getValue(thisRef: LifecycleOwner, property: KProperty<*>): Scope {
-        return if (_scope != null) _scope!!
+        return if (_scope == null) {
+            createScope()
+            return _scope ?: error("can't get Scope for $lifecycleOwner")
+        }
         else {
-            if (!thisRef.isActive()) {
-                error("can't get Scope for $lifecycleOwner - LifecycleOwner is not Active")
-            } else {
-                val koin = koinContext.get()
-                _scope = koin.getScopeOrNull(thisRef.getScopeId()) ?: createScope(koin)
-                koin.logger.debug("got scope: $_scope for $lifecycleOwner")
-                _scope!!
-            }
+            _scope ?: error("can't get Scope for $lifecycleOwner")
         }
     }
 }