+++ examples/androidx-samples/src/main/java/org/koin/sample/androidx/components/main/SimplePresenter.kt
@@ -2,4 +2,4 @@ package org.koin.sample.androidx.components.main
 
 import android.app.Activity
 
-class SimplePresenter(val activity: Activity)
+//class SimplePresenter(val activity: Activity)